+++ src/routes/auth.py
@@ -16,8 +16,7 @@ def login():
 
     if form.validate_on_submit():
         user = User.query.filter_by(username=form.username.data).first()
-        #if user and user.check_password(form.password.data):
-        if user and user.password == form.password.data:
+        if user and user.check_password(form.password.data):
             login_user(user)
             return redirect(url_for('main.index'))
         else: